# Improve texture implementation and documentation