We have a lot of great work in `pyvista.Texture` that isn't exposed via our documentation. This PR exposes the class in our documentation and:

- Improves the class documentation
- Adds `repr` to `pyvista.Texture`
- Allow init from 6 images to construct a `cubemap`. The decision was made to add it here since the `pyvista.cubemap` helper is already cluttered.
- Add in a `plot` method that works with cubemap enabled Textures.
- Support greyscale textures.
- Add rotation methods.
- Speed up array access and improve performance:

#### `main` branch

```py

>>> from pyvista import examples
>>> texture = examples.load_globe_texture()

>>> timeit texture.to_array()
659 µs ± 67.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

>>> timeit texture.plot(off_screen=True)
494 ms ± 73.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

#### This branch

```py
>>> timeit texture.image_data
48.9 µs ± 1.06 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

>>> timeit texture.plot(off_screen=True)
421 ms ± 8.88 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

### Bonus

Fixed the `topo-map.py` example.