@akaszynski 2022-06-24
I ended up going down the rabbit hole of rewriting how we output point and cell data for structured datasets. I'm going to pull this out into a separate PR, but open an issue to see if this is an acceptable change.